Refactor tuple code into separate file
