+++ soroban-sdk/src/events.rs
@@ -3,9 +3,7 @@ use core::fmt::Debug;
 
 #[cfg(doc)]
 use crate::{contracttype, Bytes, Map};
-use crate::{
-    env::internal, unwrap::UnwrapInfallible, ConversionError, Env, IntoVal, TryFromVal, Val, Vec,
-};
+use crate::{env::internal, unwrap::UnwrapInfallible, Env, IntoVal, Val, Vec};
 
 // TODO: consolidate with host::events::TOPIC_BYTES_LENGTH_LIMIT
 const TOPIC_BYTES_LENGTH_LIMIT: u32 = 32;
@@ -58,31 +56,6 @@ impl Debug for Events {
 
 pub trait Topics: IntoVal<Env, Vec<Val>> {}
 
-impl TryFromVal<Env, ()> for Vec<Val> {
-    type Error = ConversionError;
-
-    fn try_from_val(env: &Env, _v: &()) -> Result<Self, Self::Error> {
-        Ok(Vec::<Val>::new(env))
-    }
-}
-
-macro_rules! impl_topics_for_tuple {
-    ( $($typ:ident $idx:tt)* ) => {
-        impl<$($typ),*> Topics for ($($typ,)*)
-        where
-            $($typ: IntoVal<Env, Val>),*
-        {
-        }
-    };
-}
-
-// 0 topics
-impl Topics for () {}
-// 1-4 topics
-impl_topics_for_tuple! { T0 0 }
-impl_topics_for_tuple! { T0 0 T1 1 }
-impl_topics_for_tuple! { T0 0 T1 1 T2 2 }
-impl_topics_for_tuple! { T0 0 T1 1 T2 2 T3 3 }
 
 impl Events {
     #[inline(always)]
